% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMultiDiffMeth.R
\name{calculateMultiDiffMeth}
\alias{calculateMultiDiffMeth}
\title{Calculate the multiDiff Object}
\usage{
calculateMultiDiffMeth(meth, cur_design, cur_formula, num.cores = 2,
  .parallel = T, stat.method = "Wald", diff.method = "max")
}
\arguments{
\item{meth}{methBase object}

\item{cur_design}{the design matrix}

\item{cur_formula}{the design formula, including interactions}

\item{num.cores}{number of cores if running in parallel.}

\item{.parallel}{whether to run with multiple cores or not.}

\item{stat.method}{Which statistical test to use, Wald, or LLT. Defaults to Wald.}

\item{diff.method}{which method to use for effect size calculation. Defaults to "maximum"}
}
\description{
This function calculates differential methylation for multiple covariates
from a methylBase object, and an associated array and design formula, and returns a multiDiff array.
}
\keyword{multiDiff}
