% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMultiOnMulti.R
\name{filterMultiOnMulti}
\alias{filterMultiOnMulti}
\title{Filter the sites of a multiDiff array against another}
\usage{
filterMultiOnMulti(baseMulti, filtMulti, inv = F, coef, meth.thresh = 25,
  q.thresh = 0.01)
}
\arguments{
\item{baseMulti}{the base multiDiff object to be filtered}

\item{filtMulti}{the multiDiff object used for filtering.}

\item{inv}{The default behavior is to only include sites in the baseMulti
that are called as differential in filtMulti. This flag inverts the filter,
only letting through sites that weren't called. Default=F.}

\item{coef}{Which coefficient in filtMulti to use for filtering.}

\item{meth.thresh}{Effect size threshold for differential calling. Default=25.}

\item{q.thresh}{Statistical threshold for differential calling. Default=0.01}
}
\description{
Effectively does an inner join between all the sites in the baseMulti,
and the differential sites found for "coef" in filtMulti.
"inv" can be used to make it a left join instead.
on baseMulti.
}
\keyword{differential,DiffMatrix}
