% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDiffCallMatrix.R
\name{getDiffCallMatrix}
\alias{getDiffCallMatrix}
\title{Get the differential call matrix from a multiDiff array}
\usage{
getDiffCallMatrix(multiDiff, meth.thresh = 10, q.thresh = 0.01,
  long.rownames = F)
}
\arguments{
\item{multiDiff}{A multiDiff object}

\item{meth.thresh}{Effect size threshold for differential calling. Default=10.}

\item{q.thresh}{Statistical threshold for diffetential calling. Default=0.01}

\item{long.rownames}{Inserts rownames which give chr and pos. Default=F}
}
\description{
Pulls out the binary matrix of called sites for all covariates.
}
\keyword{differential,DiffCallMatrix}
