% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHeatMap.R
\name{getDiffMatrix}
\alias{getDiffMatrix}
\title{Get the differential effect matrix from a multiDiff array}
\usage{
getDiffMatrix(multiDiff)
}
\arguments{
\item{multiDiff}{A multiDiff object}
}
\description{
Pulls out the matrix of called effect size changes for all covariates.
}
\keyword{differential,DiffMatrix}
