% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethMatrix.R
\name{getMethMatrix}
\alias{getMethMatrix}
\title{Get the raw methylation matrix from a methBase object}
\usage{
getMethMatrix(mbase, long.rownames = F)
}
\arguments{
\item{mbase}{A methylBase object}

\item{long.rownames}{Inserts rownames which give chr and pos. Default=F}
}
\description{
Pulls out raw methylation values for all samples. Missing data is filled with NAs.
}
\keyword{methylBase,methylKit}
