% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDiffCallMatrix.R
\name{getSummaryDiffCall}
\alias{getSummaryDiffCall}
\title{Summarize the differential calls of a multiDiff object}
\usage{
getSummaryDiffCall(multiDiff, meth.thresh = 10, q.thresh = 0.01)
}
\arguments{
\item{multiDiff}{A multiDiff object}

\item{meth.thresh}{Effect size threshold for differential calling. Default=10.}

\item{q.thresh}{Statistical threshold for diffetential calling Default=0.01}
}
\description{
Summarizes the number of called hyper, null, and hypo, sites for all covariates.
}
\keyword{differential,SummaryDiffCallMatrix}
