% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHeatMap.R
\name{makeHeatMap}
\alias{makeHeatMap}
\title{Create a heatmap for a multiDiff object}
\usage{
makeHeatMap(multiDiff, meth.thresh = 10, q.thresh = 0.01, make.plot = T,
  cluster_cols = F, call_col = "black", plot_all_sites = F,
  long_chr = F, relabel = NA, transpose = F, fontsize = 18)
}
\arguments{
\item{multiDiff}{multiDiff array.}

\item{meth.thresh}{Effect size threshold for differential calling. Default=10.}

\item{q.thresh}{Statistical threshold for differential calling. Default=0.01.}

\item{make.plot}{Whether to make the plot. Default=T.}

\item{cluster_cols}{Whether to cluster the columns (sites). Can be memory intensive Default=F.}

\item{call_col}{Color of the called annotations. Default='black'.}

\item{plot_all_sites}{Whether to plot all sites, instead of only differential sites. Default=F.}

\item{long_chr}{Whether to use "chromosome" vs "chr" Default=F.}

\item{relabel}{Option string to relabel covariates. Default=NA.}

\item{transpose}{Whether to transpose the plot. Default=F.}

\item{fontsize}{What fontsize to use. Default=18.}
}
\description{
Allows quick generation of heatmaps for a multiDiff object.
For larger plots, png file format is recommended if sites are to be clustered, due to memory issues.
}
\keyword{heatmap}
