% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeViolinPlot.R
\name{makeViolinPlot}
\alias{makeViolinPlot}
\title{Create a violin and/or barplot for a multiDiff object}
\usage{
makeViolinPlot(multiDiff, meth.thresh = 0, q.thresh = 0.01,
  make.violin = T, make.bar = T, make.plot = T, bw_plot = F,
  relabel = NA, bar.ymax = NA, fontsize = 16)
}
\arguments{
\item{multiDiff}{multiDiff array.}

\item{meth.thresh}{Effect size threshold for differential calling. Default=0.}

\item{q.thresh}{Statistical threshold for differential calling. Default=0.01}

\item{make.violin}{Whether to make the violin plot. Default=T.}

\item{make.bar}{Whether to make the bar plot. Default=T.}

\item{bw_plot}{Whether to make the plot black and white. Default=F}

\item{relabel}{Optional string to relabel covariates. Default=NA.}

\item{bar.ymax}{Maximum of y-axis in bar plot. Default=NA}

\item{fontsize}{What fontsize to use. Default=18.}
}
\description{
Allows quick generation of violin and barplots for a multiDiff object.
}
\keyword{heatmap}
