% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_meth_sim.R
\name{run_meth_sim}
\alias{run_meth_sim}
\title{Run methylation simulation}
\usage{
run_meth_sim(base_only = F, sim_num_sites_per_cond = 1000,
  sim_read_depth = 10, sim_num_samps_per_group = 4, seed = 1,
  sim_control_region_size_multiplier = 1, sim_use_sawtooth_base = F,
  sim_mean_beta = 3, sim_beta_multiplier = c(1, -1, -1),
  sim_condition_noise = 0.01, sim_biological_noise = 0.01, run_DSS = T,
  make_confusion_matrices = T, link_func = logit, inv_link_func = ilogit,
  sim_cores = 4)
}
\arguments{
\item{base_only}{Whether to generate multDiff and DSS call objects. Will only
return a methylBase object if set to true. Default=F}

\item{sim_num_sites_per_cond}{Number of sites per condition. Default=1000}

\item{sim_read_depth}{Read depth- simulated uniformly. Default=10}

\item{sim_num_samps_per_group}{Number of samples for each of the four biological conditions. Default=4}

\item{seed}{Random seed. Default=1}

\item{sim_control_region_size_multiplier}{Allows the control region to be larger to
simulate that most of the genome should not be differential. Default=1}

\item{sim_use_sawtooth_base}{Instead of a flat baseline, simulates a sawtooth patten. Default=F.}

\item{sim_mean_beta}{Mean linear effect of the covariates. Default=3}

\item{sim_condition_noise}{Controls how much variance there is in the effect size across each region. Default=0.01}

\item{sim_biological_noise}{Simulates the biological noise across replicates. Default=0.01}

\item{run_DSS}{Whether to run DSS or not. Default =T.}

\item{link_func}{The link function. Expects inputs in [0,1]. Defaults to logit.}

\item{inv_link_func}{The inverse link function. Expected to output in [0,1], should be inverse of the link.}

\item{sim_cores}{Number of cores to use in the simulation}

\item{sim_beta_multiplier=c(1, -1, -1)}{Vector of 3 numbers allowing each covariate
and their interaction term to have different strengths. Default =c(1,-1,-1).}
}
\description{
Runs multivariate simulation of methylation for methylKit and DSS
If both the DSS and confusion matrix flags are set to false,
will return a standard multiDiff array. Otherwise will return a list of lists.
}
